refactor out mockdatabase
